requirements.txt
anyio==3.7.1
certifi==2023.7.22
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sniffio==1.3.0
